Pull most files from GitHub over https:// to protect against MITM
NLTK just uses GitHub Pages for its website (and for hosting the
various corpus packages), and GitHub offers an easy way to securely
download files over https:// rather than unencrypted (and insecure)
http://.

Swap every file download to use the appropriate https:// URL except
PanLex, which isn't stored in the repository (since it's too large).

Next step would then to have the NLTK python library use the https://
URL for the index.xml so that everything is end-to-end encrypted
and validated.